add Project and Task

+++ Repositories/Repositories/GenericRepository.cs
@@ -29,7 +29,7 @@ namespace Projet.Repositories.Repositories
             return await _context.Set<T>().ToListAsync();
         }
 
-        public async Task<T> GetByIdAsync(int id)
+        public async Task<T> GetByIdAsync(long id)
         {
             return await _context.Set<T>().FindAsync(id);
         }
